refactor(dtni-v1): consolidate dtni modules under cvs/lib/dtni/
Restructures the DTNI Step-1 layout to reduce visual clutter and align with
existing repo conventions:

- cvs/lib/{adapter_protocol,arch_detect,artifact_writer,base_adapter,
  catalog,config_loader,errors,executor,hashing,job,resource_resolver,
  run_context,substitution,topology,verdict}.py -> cvs/lib/dtni/*.py
- cvs/lib/benchmarks/ -> cvs/lib/dtni/benchmarks/
- cvs/lib/runtime/container_handle.py -> cvs/lib/dtni/container_handle.py
  (runtime/ subdir removed; only one file lived there)
- cvs/frameworks/vllm_single/adapter.py -> cvs/frameworks/vllm_single_adapter.py
  (vllm_single/ subdir removed; one file)
- cvs/tests_dtni/unittests/ -> cvs/lib/dtni/unittests/
  (matches cvs/lib/preflight/unittests/ precedent)

All imports rewritten across 19 files. No behavior changes. 79/79 DTNI
unittests pass.

Gate: PYTHONPATH=. python3 -m pytest cvs/lib/dtni/unittests/ -c /dev/null --noconftest
