+++ cvs/lib/dtni/adapter_protocol.py
