+++ cvs/lib/dtni/arch_detect.py
