+++ cvs/lib/dtni/artifact_writer.py
