+++ cvs/lib/dtni/benchmarks/projectors.py
@@ -19,7 +19,7 @@ from __future__ import annotations
 
 from typing import Any, Callable
 
-from cvs.lib.benchmarks.registry import BenchmarkSpec
+from cvs.lib.dtni.benchmarks.registry import BenchmarkSpec
 
 
 # ---- lm-eval-harness ------------------------------------------------------
