+++ cvs/lib/dtni/benchmarks/registry.py
