+++ cvs/lib/dtni/container_handle.py
