+++ cvs/lib/dtni/errors.py
