+++ cvs/lib/dtni/hashing.py
