+++ cvs/lib/dtni/substitution.py
