+++ cvs/lib/dtni/topology.py
