+++ cvs/lib/dtni/verdict.py
